Overhaul of the exporter. - Moved AWD, OBJ, and Serialization export methods into classes and created extensible framework for general exporting. - Removed superfluous serializeOBJ methods as they are redundant. - Added code examples and class documentation to AWD parsing and other related classes. - Updated class documentation for some of the various classes I have written.
