+++ src/rajawali/util/MeshExporter.java
@@ -1,72 +1,128 @@
 package rajawali.util;
 
-import java.io.BufferedWriter;
 import java.io.File;
-import java.io.FileOutputStream;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.ObjectOutputStream;
-import java.io.Writer;
-import java.nio.Buffer;
-import java.nio.FloatBuffer;
-import java.nio.IntBuffer;
-import java.nio.ShortBuffer;
-import java.util.zip.GZIPOutputStream;
 
 import rajawali.BaseObject3D;
-import rajawali.Geometry3D;
-import rajawali.SerializedObject3D;
-import rajawali.animation.mesh.VertexAnimationFrame;
-import rajawali.animation.mesh.VertexAnimationObject3D;
-import rajawali.materials.textures.TextureManager;
-import rajawali.parser.ObjParser;
-import android.content.Context;
+import rajawali.util.exporter.AExporter;
 import android.os.Environment;
 
+/**
+ * Utility class for exporting models in various types. The exporter is flexible and should permit support of
+ * essentially any file type. In the future the exporter will be updated to support scenes for AWD exporting and any
+ * other model types that may support it.
+ * <p>
+ * 
+ * <b>Example AWD exporting</b>
+ * 
+ * <code><pre>
+ * 	BaseObject3D cube = new Cube(1f);
+ * 	MeshExporter exporter = new MeshExporter(cube);
+ * 	exporter.export(&quot;cube&quot;, AwdExporter.class);
+ * </pre></code>
+ * 
+ * 
+ * @author Ian Thomas (toxicbakery@gmail.com)
+ * 
+ */
 public class MeshExporter {
 
 	private BaseObject3D mObject;
 	private String mFileName;
-	private boolean mCompressed;
 	private File mExportDir = null;
-
-	public enum ExportType {
-		AWD,
-		SERIALIZED,
-		OBJ
-	}
+	private boolean mAppendExtension;
 
 	public MeshExporter(BaseObject3D objectToExport) {
 		mObject = objectToExport;
+		mAppendExtension = true;
 	}
 
-	public void setExportDirectory(File exportDir) {
-		// The path's validity is the user's responsibility.
-		// Any problems are taken care of by the try blocks in the private export methods.
-		mExportDir = exportDir;
+	/**
+	 * Export the model using the set exporter without compression. The filename is appended with the extension defined
+	 * by the exporter if enabled.
+	 * 
+	 * @param fileName
+	 * @param aExporter
+	 * @throws ExporterException
+	 */
+	public void export(String fileName, Class<? extends AExporter> aExporter) throws ExporterException {
+		export(fileName, aExporter, false);
 	}
 
-	public void export(String fileName, ExportType type) {
-		export(fileName, type, false);
-	}
+	/**
+	 * Export the model using the set exporter and compression. The filename is appended with the extension defined by
+	 * the exporter if enabled.
+	 * 
+	 * @param fileName
+	 * @param aExporter
+	 * @param compressed
+	 * @throws ExporterException
+	 */
+	public void export(String fileName, Class<? extends AExporter> aExporter, boolean compressed)
+			throws ExporterException {
+		try {
+			// Construct the parser
+			final AExporter parser = (AExporter) Class.forName(aExporter.getName()).getConstructor()
+					.newInstance();
 
-	public void export(String fileName, ExportType type, boolean compressed) {
-		mFileName = fileName;
-		mCompressed = compressed;
-		switch (type) {
-		case AWD:
-			exportToAwd();
-			break;
-		case SERIALIZED:
-			exportToSerialized();
-			break;
-		case OBJ:
-			exportToObj();
-			break;
+			// Determine the output filename. Append the extension if flagged to do so.
+			mFileName = mAppendExtension ? fileName + "." + parser.getExtension() : fileName;
+
+			// Verify that the file exists or alternately can be written to.
+			final File exportFile = getExportFile();
+
+			// Create the file if it does not exist
+			if (!exportFile.exists()) {
+				exportFile.getParentFile().mkdirs();
+				exportFile.createNewFile();
+			}
+
+			// Verify the file can be written to
+			if (!exportFile.canWrite())
+				throw new ExporterException(mFileName + " can not be written to.");
+
+			// Configure the parser
+			parser.setExporter(this);
+			parser.setExportFile(getExportFile());
+			parser.setExportModel(mObject);
+			parser.setCompressed(compressed);
+			configureExporter(parser);
+			parser.export();
+		} catch (Exception e) {
+			throw new ExporterException(e);
 		}
 	}
 
-	private File getExportFile() {
+	/**
+	 * Determine if the extension defined by the exporter should be automatically appended. This is true by default.
+	 * 
+	 * @param flag
+	 */
+	public void setAppendExtension(boolean flag) {
+		mAppendExtension = flag;
+	}
+
+	/**
+	 * Directory to export the model to.
+	 * 
+	 * @param exportDir
+	 */
+	public void setExportDirectory(File exportDir) {
+		mExportDir = exportDir;
+	}
+
+	/**
+	 * Method available for override to additionally configure a parser.
+	 * 
+	 * @param exporter
+	 */
+	protected void configureExporter(AExporter exporter) {}
+
+	/**
+	 * The file to be exported to.
+	 * 
+	 * @return
+	 */
+	protected File getExportFile() {
 		File path = mExportDir;
 		if (mExportDir == null)
 			path = Environment.getExternalStorageDirectory();
@@ -74,288 +130,22 @@ public class MeshExporter {
 		return new File(path, mFileName);
 	}
 
-	private void exportToAwd() {
-		final File f = getExportFile();
+	public static final class ExporterException extends RuntimeException {
 
-		RajLog.d("Exporting as AWD2 file");
-		RajLog.d("Writing to " + f.getAbsolutePath());
+		private static final long serialVersionUID = 1L;
 
-		try {
-			// Create a buffered LittleEndianOutputStream for writing
-			final FileOutputStream fos = new FileOutputStream(f);
-			final LittleEndianOutputStream los = new LittleEndianOutputStream(fos);
-
-			// Write the magic string "AWD"
-			los.writeByte(65);
-			los.writeByte(87);
-			los.writeByte(68);
-
-			// Write the document header data
-			los.write(2);
-			los.write(0);
-			los.writeShort(0);
-			los.write(0);
-
-			// Write the file size, this should really print the total size post header but it would be difficult to
-			// determine the total size before writing.
-			los.writeInt(0);
-
-			// Start the Triangle block
-			los.writeInt(0);// ID
-			los.write(0);// Namespace
-			los.write(1);// Type
-			los.write(0);// Flags
-			los.writeInt(0); // Length
-
-			// Write the name
-			los.writeShort(mObject.getName() == null ? 0 : mObject.getName().length());
-			if (mObject.getName() != null)
-				los.write(mObject.getName().getBytes());
-
-			// Write the sub geometry count
-			los.writeShort(1);
-
-			// Write the properties
-			los.writeInt(0);
-
-			final Geometry3D geom = mObject.getGeometry();
-
-			// Write the sub mesh length
-			los.writeInt(awdGetGeomLength(geom));
-
-			// Write the properties
-			los.writeInt(0);
-
-			// Write the geometry
-			writeAwdAttributeList(los, 1, 0, geom.getVertices());
-			writeAwdAttributeList(los, 2, 0, geom.getIndices());
-			writeAwdAttributeList(los, 3, 0, geom.getTextureCoords());
-			writeAwdAttributeList(los, 4, 0, geom.getNormals());
-
-			// Close (fos has to be closed separate of the los or writing does not close properly)
-			fos.close();
-			los.close();
-
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-
-	private int awdGetGeomLength(Geometry3D geom) {
-		return 24 + (geom.getNumIndices() * 2) + (geom.getNumVertices() * 4) + (geom.getNormals().limit() * 4)
-				+ (geom.getTextureCoords().limit() * 4);
-	}
-
-	private void writeAwdAttributeList(LittleEndianOutputStream los, int type, int typeF, Buffer data)
-			throws Exception {
-		// Type
-		los.writeByte(type);
-		los.writeByte(typeF);
-
-		if (data instanceof IntBuffer) {
-			// Length of mesh data
-			los.writeInt(data.limit() * 2);
-
-			final IntBuffer buf = (IntBuffer) data;
-			for (int i = 0, j = data.limit(); i < j; i++)
-				los.writeShort(buf.get());
-		} else if (data instanceof ShortBuffer) {
-			// Length of mesh data
-			los.writeInt(data.limit() * 2);
-
-			final ShortBuffer buf = (ShortBuffer) data;
-			for (int i = 0, j = data.limit(); i < j; i++)
-				los.writeShort(buf.get());
-		} else if (data instanceof FloatBuffer) {
-			// Length of mesh data
-			los.writeInt(data.limit() * 4);
-
-			final FloatBuffer buf = (FloatBuffer) data;
-			for (int i = 0, j = data.limit(); i < j; i++)
-				los.writeFloat(buf.get());
-		}
-	}
-
-	private void exportToObj() {
-		RajLog.d("Exporting " + mObject.getName() + " as .obj file");
-		Geometry3D g = mObject.getGeometry();
-		StringBuffer sb = new StringBuffer(9000);
-		File f = getExportFile();
-		BufferedWriter bw = null;
-
-		try {
-			bw = new BufferedWriter(new FileWriter(getExportFile()));
-		} catch (IOException e) {
-			e.printStackTrace();
+		public ExporterException(final String msg) {
+			super(msg);
 		}
 
-		sb.append("# Exported by Rajawali 3D Engine for Android\n");
-		sb.append("o ");
-		sb.append(mObject.getName());
-		sb.append("\n");
-
-		for (int i = 0; i < g.getVertices().capacity(); i += 3) {
-			sb.append("v ");
-			sb.append(g.getVertices().get(i));
-			sb.append(" ");
-			sb.append(g.getVertices().get(i + 1));
-			sb.append(" ");
-			sb.append(g.getVertices().get(i + 2));
-			sb.append("\n");
-			bufferStringWriting(sb, bw);
+		public ExporterException(Throwable e) {
+			super(e);
 		}
 
-		sb.append("\n");
-
-		for (int i = 0; i < g.getTextureCoords().capacity(); i += 2) {
-			sb.append("vt ");
-			sb.append(g.getTextureCoords().get(i));
-			sb.append(" ");
-			sb.append(g.getTextureCoords().get(i + 1));
-			sb.append("\n");
-			bufferStringWriting(sb, bw);
-		}
-
-		sb.append("\n");
-
-		for (int i = 0; i < g.getNormals().capacity(); i += 3) {
-			sb.append("vn ");
-			sb.append(g.getNormals().get(i));
-			sb.append(" ");
-			sb.append(g.getNormals().get(i + 1));
-			sb.append(" ");
-			sb.append(g.getNormals().get(i + 2));
-			sb.append("\n");
-			bufferStringWriting(sb, bw);
-		}
-
-		sb.append("\n");
-
-		boolean isIntBuffer = g.getIndices() instanceof IntBuffer;
-
-		for (int i = 0; i < g.getIndices().capacity(); i++) {
-			if (i % 3 == 0)
-				sb.append("\nf ");
-			int index = isIntBuffer ? ((IntBuffer) g.getIndices()).get(i) + 1
-					: ((ShortBuffer) g.getIndices()).get(i) + 1;
-			sb.append(index);
-			sb.append("/");
-			sb.append(index);
-			sb.append("/");
-			sb.append(index);
-			sb.append(" ");
-			bufferStringWriting(sb, bw);
-		}
-
-		try {
-			// Write any remaining data to the file
-			bw.append(sb.toString());
-			bw.flush();
-			bw.close();
-
-			RajLog.d(".obj export successful: " + f.getCanonicalPath());
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-	}
-
-	/**
-	 * Make sure this line is in your AndroidManifer.xml file, under <manifest>: <uses-permission
-	 * android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
-	 */
-	private void exportToSerialized() {
-		FileOutputStream fos;
-		try {
-			File f = getExportFile();
-			fos = new FileOutputStream(f);
-			ObjectOutputStream os = null;
-			if (mCompressed) {
-				GZIPOutputStream gz = new GZIPOutputStream(fos);
-				os = new ObjectOutputStream(gz);
-			} else {
-				os = new ObjectOutputStream(fos);
-			}
-
-			SerializedObject3D ser = mObject.toSerializedObject3D();
-
-			if (mObject instanceof VertexAnimationObject3D) {
-				VertexAnimationObject3D o = (VertexAnimationObject3D) mObject;
-				int numFrames = o.getNumFrames();
-				float[][] vs = new float[numFrames][];
-				float[][] ns = new float[numFrames][];
-				String[] frameNames = new String[numFrames];
-
-				for (int i = 0; i < numFrames; ++i) {
-					VertexAnimationFrame frame = (VertexAnimationFrame) o.getFrame(i);
-					Geometry3D geom = frame.getGeometry();
-					float[] v = new float[geom.getVertices().limit()];
-					geom.getVertices().get(v);
-					float[] n = new float[geom.getNormals().limit()];
-					geom.getNormals().get(n);
-					vs[i] = v;
-					ns[i] = n;
-					frameNames[i] = frame.getName();
-				}
-
-				ser.setFrameVertices(vs);
-				ser.setFrameNormals(ns);
-				ser.setFrameNames(frameNames);
-			}
-
-			os.writeObject(ser);
-			os.close();
-			RajLog.i("Successfully serialized " + mFileName + " to " + f.getCanonicalPath());
-		} catch (Exception e) {
-			RajLog.e("Serializing " + mFileName + " was unsuccessfull.");
-			e.printStackTrace();
+		public ExporterException(final String msg, Throwable e) {
+			super(msg, e);
 		}
 
 	}
 
-	/**
-	 * Helper method for writing chunks of data to the given writer. If data is written the passed StringBuffer will be
-	 * emptied.
-	 * 
-	 * @param stringBuilder
-	 * @param writer
-	 */
-	private void bufferStringWriting(StringBuffer stringBuilder, Writer writer) {
-		if (stringBuilder.length() >= 8192) {
-			try {
-				writer.write(stringBuilder.toString());
-				stringBuilder.delete(0, stringBuilder.length());
-			} catch (IOException e) {
-				e.printStackTrace();
-			}
-		}
-	}
-
-	public static void serializeObj(Context context, TextureManager textureManager, int resourceId, String outputName) {
-		serializeObj(context, textureManager, resourceId, outputName, false, null);
-	}
-
-	public static void serializeObj(Context context, TextureManager textureManager, int resourceId, String outputName,
-			Boolean compress) {
-		serializeObj(context, textureManager, resourceId, outputName, compress, null);
-	}
-
-	public static void serializeObj(Context context, TextureManager textureManager, int resourceId, String outputName,
-			File exportDir) {
-		serializeObj(context, textureManager, resourceId, outputName, false, exportDir);
-	}
-
-	public static void serializeObj(Context context, TextureManager textureManager, int resourceId, String outputName,
-			Boolean compress, File exportDir) {
-		final ObjParser objParser = new ObjParser(context.getResources(), textureManager, resourceId);
-		try {
-			objParser.parse();
-			final BaseObject3D obj = objParser.getParsedObject();
-			final MeshExporter exporter = new MeshExporter(obj);
-			exporter.setExportDirectory(exportDir);
-			exporter.export(outputName, ExportType.SERIALIZED, compress);
-		} catch (Exception e) {
-			RajLog.e("Failed to serialize obj: " + e.getMessage());
-			e.printStackTrace();
-		}
-	}
 }
